+++ jest.config.js
@@ -1,3 +1,5 @@
 module.exports = {
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testPathIgnorePatterns: ['/e2e/'],
 };